requirements.txt
streamlit
pandas
numpy
openpyxl
plotly
pydantic
python-dotenv
openai
pytest